moved methods from PgpMsg to MsgBlockParser
